jk/reference updater (#1957)
* create separate ReferenceUpdater module

* remove unnecessary imports from ReferenceTests

* remove html

* move and edit readme

* mention ctrl c
